requirements.txt
anthropic>=0.75.0
feedparser>=6.0.12
python-dotenv>=1.2.1
requests>=2.32.5
schedule>=1.2.2
